+++ CHANGELOG.md
@@ -7,4 +7,5 @@ CHANGELOG
 
 - fixed issue #1 KL
 - Fixed issue #2; Added Last Controller
-- fixed issues #4; added wildcard for lock/em/up KL 
+- fixed issues #4; added wildcard for lock/em/up KL
+- Fixed issues #6; Dunno route - Jordan Hamade